requirements.txt
Flask==0.10.1
Flask-Admin==1.0.8
Flask-Bootstrap==3.2.0.2
Flask-Login==0.2.11
Flask-Mail==0.9.0
Flask-Migrate==1.2.0
Flask-Misaka==0.3.0
Flask-Principal==0.4.0
Flask-RESTful==0.2.12
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask-WTF==0.10.2
Jinja2==2.7.3
Mako==1.0.0
MarkupSafe==0.23
Pillow==2.5.3
PyMySQL==0.6.2
SQLAlchemy==0.9.7
WTForms==2.0.1
Werkzeug==0.9.6
alembic==0.6.7
aniso8601==0.83
argparse==1.2.1
blinker==1.3
distribute==0.6.24
itsdangerous==0.24
marshmallow==0.7.0
misaka==1.0.2
passlib==1.6.2
py-bcrypt==0.4
pytz==2014.7
six==1.8.0
wsgiref==0.1.2
